backend/requirements.txt
absl-py==2.3.1
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
astunparse==1.6.3
attrs==22.1.0
azure-cognitiveservices-speech==1.44.0
azure-core==1.34.0
azure-digitaltwins-core==1.2.0
azure-identity==1.23.0
azure-kusto-data==5.0.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
cffi==1.15.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.2.0
cryptography==45.0.4
cycler==0.12.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
entrypoints==0.4
executing==1.2.0
fastapi==0.116.1
fastjsonschema==2.16.2
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.49.0
fsspec==2025.3.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.2.1
grpcio==1.73.1
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
idna==3.4
ijson==3.4.0
importlib_resources==6.5.2
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isodate==0.7.2
jedi==0.18.2
Jinja2==3.1.2
jiter==0.10.0
joblib==1.3.2
jsonschema==4.17.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==3.10.0
keyboard==0.13.5
kiwisolver==1.4.5
libclang==18.1.1
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
ml_dtypes==0.5.1
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
namex==0.1.0
narwhals==1.46.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.4.2
nibabel==5.3.2
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.26.4
oauthlib==3.3.1
openai==1.88.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
optuna==4.3.0
packaging==21.3
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
platformdirs==2.5.4
playsound==1.3.0
prometheus-client==0.15.0
prompt-toolkit==3.0.33
protobuf==5.29.5
psutil==5.9.4
pure-eval==0.2.2
pyarrow==20.0.0
pycparser==2.21
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.13.0
PyJWT==2.10.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2024.1
pywin32==305
pywinpty==2.0.9
PyYAML==6.0.2
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.0
six==1.16.0
skorch==1.1.0
smmap==5.0.2
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.40
stack-data==0.6.1
starlette==0.47.1
streamlit==1.46.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
terminado==0.17.0
threadpoolctl==3.3.0
tinycss2==1.2.1
toml==0.10.2
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.2
tqdm==4.67.1
traitlets==5.5.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2024.1
urllib3==2.4.0
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==3.1.3
widgetsnbextension==4.0.3
wrapt==1.17.2
